feat: more api endpoints

+++ shared/core/src/main/kotlin/space/astro/shared/core/properties/bot/BotEndpointProperties.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 @ConfigurationProperties("bot.endpoint")
 data class BotEndpointProperties(
     val localhost: Boolean,
-    val port: Long,
+    val port: Int,
     val podName: String,
     val serviceName: String,
     val namespace: String,
